// app/internal/renderer/worker-postable-generated.ts
// Auto generated by worker-postable
interface ITimeline {
    totalTime: number;
    currentTime: number;
    tracks: Array<ITrack>;
}
interface ITrackItem {
    startTime: number;
    endTime: number;
    next: ITrackItem;
    prev: ITrackItem;
}
interface ITrack {
    trackItems: Set<ITrackItem>;
}

export { ITimeline as ITimeline, ITrackItem as ITrackItem, ITrack as ITrack };